+++ tests/js/app-tests.js
@@ -124,6 +124,7 @@ module.exports = {
         let credentials = Realm.Credentials.anonymous();
         let user = await app.logIn(credentials);
         TestCase.assertInstanceOf(user, Realm.User);
+        TestCase.assertNotNull(user.deviceId);
         await user.logOut();
     },